LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.14-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token 'instanceof'
S7.5.2_A1.14

PASS Expected parsing failure

TEST COMPLETE
